+++ example/get_android_dependencies.sh
@@ -1,9 +1,14 @@
 #!/bin/sh
+
+AneVersion="1.6.0"
+PlayerServicesVersion="15.0.1"
+SupportV4Version="27.1.0"
+
 wget -O android_dependencies/com.tuarua.frekotlin.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/kotlin/com.tuarua.frekotlin.ane?raw=true
 wget -O android_dependencies/org.greenrobot.eventbus-3.0.0.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/misc/org.greenrobot.eventbus-3.0.0.ane?raw=true
 wget -O android_dependencies/com.google.code.gson.gson-2.8.1.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/misc/com.google.code.gson.gson-2.8.1.ane?raw=true
-wget -O android_dependencies/com.android.support.support-v4-27.1.0.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/support/com.android.support.support-v4-27.1.0.ane?raw=true
-wget -O android_dependencies/com.google.android.gms.play-services-base-12.0.1.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-base-12.0.1.ane?raw=true
-wget -O android_dependencies/com.google.android.gms.play-services-location-12.0.1.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-#location-12.0.1.ane?raw=true
-wget -O android_dependencies/com.google.android.gms.play-services-maps-12.0.1.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-maps-12.0.1.ane?#raw=true
-wget -O ../native_extension/ane/GoogleMapsANE.ane https://github.com/tuarua/Google-Maps-ANE/releases/download/1.6.0/GoogleMapsANE.ane?raw=true
+wget -O android_dependencies/com.android.support.support-v4-$SupportV4Version.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/support/com.android.support.support-v4-$SupportV4Version.ane?raw=true
+wget -O android_dependencies/com.google.android.gms.play-services-base-$PlayerServicesVersion.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-base-$PlayerServicesVersion.ane?raw=true
+wget -O android_dependencies/com.google.android.gms.play-services-location-$PlayerServicesVersion.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-location-$PlayerServicesVersion.ane?raw=true
+wget -O android_dependencies/com.google.android.gms.play-services-maps-$PlayerServicesVersion.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-maps-$PlayerServicesVersion.ane?#raw=true
+wget -O ../native_extension/ane/GoogleMapsANE.ane https://github.com/tuarua/Google-Maps-ANE/releases/download/$AneVersion/GoogleMapsANE.ane?raw=true