+++ example/get_ios_dependencies.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
+AneVersion="1.6.0"
+
 rm -r ios_dependencies/device
 rm -r ios_dependencies/simulator
 
@@ -7,9 +9,9 @@ wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/2.4.0/ios_depende
 unzip -u -o ios_dependencies.zip
 rm ios_dependencies.zip
 
-wget https://github.com/tuarua/Google-Maps-ANE/releases/download/1.6.0/ios_dependencies.zip
+wget https://github.com/tuarua/Google-Maps-ANE/releases/download/$AneVersion/ios_dependencies.zip
 unzip -u -o ios_dependencies.zip
 rm ios_dependencies.zip
 
-wget -O ../native_extension/ane/GoogleMapsANE.ane https://github.com/tuarua/Google-Maps-ANE/releases/download/1.6.0/GoogleMapsANE.ane?raw=true
+wget -O ../native_extension/ane/GoogleMapsANE.ane https://github.com/tuarua/Google-Maps-ANE/releases/download/$AneVersion/GoogleMapsANE.ane?raw=true