+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 1.6.0
 - AND: Updated to FreKotlin 1.3.0
-- AND: Updated to Android Play Services to 12.0.1
+- AND: Updated to Android Play Services to 15.0.1
 - iOS: Updated to Google Maps SDK 2.7
 
 ### 1.5.0